third_party/blink/web_tests/platform/generic/fast/events/constructors/speech-recognition-event-constructor-expected.txt

This tests the constructor for the SpeechRecognitionEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new webkitSpeechRecognitionEvent('eventType').bubbles is false
PASS new webkitSpeechRecognitionEvent('eventType').cancelable is false
PASS new webkitSpeechRecognitionEvent('eventType').results is null
PASS new webkitSpeechRecognitionEvent('eventType').resultIndex is 0
PASS new webkitSpeechRecognitionEvent('eventType', { bubbles: false }).bubbles is false
PASS new webkitSpeechRecognitionEvent('eventType', { bubbles: true }).bubbles is true
PASS new webkitSpeechRecognitionEvent('eventType', { cancelable: false }).cancelable is false
PASS new webkitSpeechRecognitionEvent('eventType', { cancelable: true }).cancelable is true
PASS new webkitSpeechRecognitionEvent('eventType', { resultIndex: 42 }).resultIndex is 42
PASS successfullyParsed is true

TEST COMPLETE
